pyproject.toml
[tool.poetry]
name = "fall-detection-ui"
version = "0.1.0"
description = ""
authors = ["Vincent Jia <vincejia@gmail.com>"]
readme = "README.md"
packages = [{include = "fall_detection_ui", from = "src"}]

[tool.poetry.dependencies]
python = ">=3.10,<3.11"
streamlit = "^1.39.0"
streamlit-webrtc = "^0.47.9"
opencv-python-headless = "^4.10.0.84"
twilio = "^9.3.0"
loguru = "^0.7.2"
pafy = "^0.5.5"
youtube-dl = "^2021.12.17"
pytube = "^15.0.0"
pytubefix = "^6.16.2"
pip = "^24.2"
tensorflow-io-gcs-filesystem = "0.27.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[[tool.poetry.source]]
name = "pypi"
priority = "primary"